requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
pathlib2>=2.3.0; python_version < '3.4' 